feat: Brand 관련 컴포저블 함수 위치 변경
