+++ app/src/main/java/nextstep/payments/model/Card.kt
@@ -1,5 +1,10 @@
 package nextstep.payments.model
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import nextstep.payments.R
+
 data class Card(
     val id: Long = 0,
     val brand: Brand,
@@ -20,3 +25,34 @@ enum class Brand {
     HANA,
     KB,
 }
+
+@Composable
+fun Brand.toColor(): Color =
+    when (this) {
+        Brand.NONE -> Color(0xFF333333)
+        Brand.BC -> Color(0xFFF04651)
+        Brand.SHINHAN -> Color(0xFF0248f8)
+        Brand.KAKAO_BANK -> Color(0xFFfce237)
+        Brand.HYUNDAI -> Color.Black
+        Brand.WOORI -> Color(0xFF2870bd)
+        Brand.LOTTE -> Color(0xFFd83528)
+        Brand.HANA -> Color(0xFF348685)
+        Brand.KB -> Color(0xFF695F54)
+    }
+
+@Composable
+fun Brand.toName(): String =
+    stringResource(
+        id =
+            when (this) {
+                Brand.NONE -> R.string.brand_none
+                Brand.BC -> R.string.brand_bc
+                Brand.SHINHAN -> R.string.brand_shinhan
+                Brand.KAKAO_BANK -> R.string.brand_kakao_bank
+                Brand.HYUNDAI -> R.string.brand_hyundai
+                Brand.WOORI -> R.string.brand_woori
+                Brand.LOTTE -> R.string.brand_lotte
+                Brand.HANA -> R.string.brand_hana
+                Brand.KB -> R.string.brand_kb
+            },
+    )
